Fixes #3: introduce $from and @share
using @type share, a namespace can stream logs to another namespace.
On the receiving end, the target namespace can intercept logs using:
<label @$from(source_namespace)>
</label>

Except for the $from macro, it looks like the log event originated from
the same namespace as the ** syntax and $thisns macro works. However,
the kubernetes.namespace_name field still points to the original
namespace of the producing pod.

This is a big change. It introduces a lot of changes at the Processor
interface and implementation.