Add virtual filter to  detect exceptions
This patch referes to #9.

Uses https://github.com/GoogleCloudPlatform/fluent-plugin-detect-exceptions

A single <filter> of type detect_exceptions is rewritten into two
<match> directives:
 * the first detects exceptions and rewrites the tag to {ID}._proc.${tag}
 * the second strips the {ID} prefix

At all places where a tag is generated (based on $labels, $thisns or **)
a second item "_proc.{original}" is now appended is now appended so that
both abc and _proc.abc are caught. As a results both `abc` and
`_proc.abc` are caught in a user's <match> or <filter> directives.