A truncating syslog plugin to handle too long tags
This addresses #11.

Tags have a max length of 32 as per RFC 5424. To be able to send to
remote syslogs, the tags needs to be truncated by the plugin.

This patch introduces a `truncating_remote_syslog` plugin with exactly
the same interface as the `remote_syslog` but it strips part of the
fluentd tag until it fits into 32 bytes. It tries to keep the container
name and the namespace as intact as possible. For example:

demo.test.test                          -> demo.test.test
kube.demo.nginx-65899c769f-5zj6d.nginx  -> demo.nginx-65899c769f-5zj*.nginx
kube.demo.test.nginx11111111._lablels.* -> demo.test.nginx11111111