Extreme validator by means of 'just_exit' plugin
Insread of just --dry-running the config for a namespace, a normal
fluentd process is now started. A <source> of type @just_exit is added
to the expanded user config which just does exit(0). If any plug-in
(filter or output) is misconfigured then fluentd exits immediately after
encountering the error.

This helps ignore namespaces with misconfigured sensitive plugins like
S3 or logzio - or any other plugin that requires credentials to work.

See #15 for more info.