Add new Datasource based on Kubernetes Informers
Within the Go API client, the Kubernetes project has created tools for
keeping local references to API resources up to date with the server
state. By creating an Informer, the kubernetes API client can perform an
intial sync of objects we are interested in and then watch for updates
to those objects instead of polling the server on a regular interval.

Changing the access pattern of the config reloader is important for
scaling kubernetes clusters. If the config reloader is set to poll for
updates on a short interval (such as 5 seconds) it can generate quite
a significant amount of traffic on the API server even in relatively
modest clusters composed of a couple dozen compute nodes. This increase
in request volume does not provide any value to the config reloader, and
is a potentially expensive side-effect on AWS if some nodes are in a
different availability zone from the kubernetes API server(s).

By creating this new Datasource, the config reloader can be configured
with an arbitrarily short update interval without having to worry about
drowning the kubernetes API server in a flood of pointless requests. The
Informer datasource maintains watches on important API objects in a
separate goroutine, and will make changes available to configreloader
updates very quickly, usually on the update immediately following the
change in the upstream API.

The new Datasource has been designed to be able to replace both existing
Kubernetes Datasources, but has only been tested against the behavior of
the Kubernetes Multimap Datasource.

Signed-off-by: Richard Jackson <richard.jackson@55places.com>