Replace usage of existing Kubernetes Datasources
Update the controller to only use the new Informer-based datasource to
replace both existing kubernetes datasource implementations.

Note: This change has been tested in multiple production-quality
clusters for several weeks, but only in a configuration that formerly
used the Multimap datasource. Only cursory testing has been done to
verify that the new datasource is compatible with the default Kubernetes
datasource.

Signed-off-by: Richard Jackson <richard.jackson@55places.com>