Add new processor to support tag expansion
A valid Fluentd configuration file allows to specify match patterns of
the type "k.{a,b}.**" that are expanded to match both tag "k.a.**" and
tag "k.b.**" and process them with the same set of rules. Furthermore,
several tags can be listed in the match pattern separated by a space to
provide an expression that will match any of the tags. Supporting such
tag expansion allows to consistently reduce the number of equivalent
rules that need to be provided in the configuration file, especially
for complex systems that contain several log parsing pipelines. Finally,
support for such feature also reduces the risk for configuration errors
typical of the copy-paste pattern that needs to be adopted when such
expansion is performed by the user.

By adding this new processor to the default processing pipeline, the config
reloader can emulate the behaviour of Fluentd and perform tag expansion
programmatically, in the early stages of the pipeline. Therefore, the
other processors will receive in input a configuration equivalent to the
one specified by the user, but in the expected expanded format. However,
the user will be able to take advantage of the tag expansion feature to
reduce the number of rules that need to be provided and demanding the
copy-paste to the config reloader, thus avoiding potential errors.

This processor supports both the "{a,b}" pattern and the "a b" pattern.
As an example, this match pattern
'<match kube.test.{x,y,z}.** kube.test.p.{a,b}'
will be expanded in the following separate match patterns
'<match kube.test.x.**>'
'<match kube.test.y.**>'
'<match kube.test.z.**>'
'<match kube.test.p.a.**>'
'<match kube.test.p.b.**>'
and the contents of the original match statement will be cloned in each
of the expanded match statements.

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>