Fix labels not being rewritten in copy plugin
This small enhancement allows the "relabel" processor to normalize labels
also when the "relabel" plugin is not directly placed inside a match
directive but rather placed in a store directive. This is a valid config
that could derive from the necessity of demultiplexing logs to several
pipelines that still need to perform additional processing before
reaching the output.

This enhancement fixes issue vmware/kube-fluentd-operator#56

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>