Add new processor to allow controlled tag rewriting
The plugin "rewrite_tag_filter" provided alongside fluentd allows conditional
tag rewriting to perform different processing pipelines based on log contents.
This same behavior is not achievable without it unless all logs are copied
to different output labels and then filtered through the grep plugin which
however introduces considerable overhead.
Clearly, however, the usage of this plugin needs to be constrained as it
allows to break the namespace isolation enforced by the config-reloader.

This enhancement introduces a new processor that allows to specify a new
virtual output plugin, "unique_rewrite_tag", which behaves exactly as the
"rewrite_tag_filter" (and has the same syntax and supported options), but
requires the new tag to be specified using the macro $tag(new_tag).
In order to match logs that are emitted by such output plugin it is
sufficient to specify the same macro in match and filter directives.
The processor will internally handle such configuration and ensure
that the rewritten tags are unique with respect to other namespaces and
thus will avoid breaking isolation while allowing conditional rewriting.

As an example, it might be necessary to divide logs based on the severity
which is the first word printed in the message from the app, process them
differently and then send them to the same output. This can be achieved as
follows:

<match kube.ns.**>
  @type unique_rewrite_tag
  <rule>
    key message
    pattern ^ERROR
    tag $tag(notifications.error)
  </rule>
  <rule>
    key message
    pattern ^FATAL
    tag $tag(notifications.fatal)
  </rule>
</match>

<filter $tag(notifications.fatal)>
  ...(perform some extra processing)
</filter>

<filter $tag(notifications.error)>
  ...(perform different processing)
</filter>

<match $tag(notifications.**)>
  ...(common output plugin)
</match>

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>