Fix namespace fluentd status annotation update
This fix addresses the case in which a processor returns an error during
the execution of the processing phase. If the namespace was previously
empty (non-configured) and no new prep configs have been added, the
previous hash and new hash match and therefore the error message is not
patched on the namespace annotation, making it hard for a user to
understand why the new configuration is not being deployed.

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>