Fix $label processor ignoring errors and not accepting "/"
This fix addresses two errors in the processor that manages the $label
macro.

The first error occurs when a bad label is used in the configurations.
The processor detects the bad label but fails to report the error and
stop the processing, therefore the unmodified tag containing the $label
macro is delivered to the fluentd config.

The second issue is the failure of the processor to accept as valid
labels containing the '/' character, which is however not only valid,
but encouraged for example by helm as a best practice for the
standard labels.

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>