Fix $labels processor rewriting all tags
The labels processor should rewrite only tags that pertain to its domain
and start with '$labels'. Other tags should not be rewritten to avoid
changing the logs that get captured by them.

Given that the '_labels.*.*' pattern is added at the end of the tags,
all other matching expressions from filter and match directives, will
still match those tags as long as they terminate with the '**' wildcard.

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>