Decouple datasource from K8S resource used to store fluentD configs
This commit introduces an interface to decouple the type of Kubernetes
resource used to store the fluentD configs from the K8S datasource.

This enhancement should bring no changes in behavior but it should
set the stage for allowing different types of k8s resources to hold
fluentD configs.

This is in particular needed to simplify the creation of a CRD to
specify a dedicated resource to hold fluentD configs.

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>