Add 'crd' datasource to exploit custom k8s resources
The Kubernetes world is evolving around the concept of operators and
custom resources to manage the various aspects and behaviors of
components.

This commit introduces a CRD for representing fluentd configurations,
to move away from the generic ConfigMap. This allows to have a
dedicated resource for fluentd configurations, thus allowing to
easily obtain fluentd configurations by querying the `fluentdconfig`
resource, or create new configurations for applications by simply
attaching a `fluentdconfig` resource to the application specs.

Furthermore, the use of CRDs should benefit the efficiency of
the config-reloader. In fact, especially in clusters with a
lot of applications that utilize ConfigMaps dynamically (for
example for leader election purposes), a lot of updates happen
to ConfigMaps, which are delivered to the config-reloader which
has to realize they are not meant for it and discard them. Using
CRDs, the config-reloader can register with the API Server to
watch only `fluentdconfig` resources, thus being notified only
when really necessary.

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>