Make the admin namespace configurable
The config-reloader assumes that the admin namespace in which configs
get defined without validation and/or processing and in which virtual
plugins get defined is the "kube-system" namespace. While this assumption
is reasonable in general, there might be instances in which the logging
infrastructure is installed in a dedicated namespace and it would be
cleaner and more manageable to be able to install virtual plugins and
global fluentd configs in that same namespace. Furthermore, it might
be the case that the kube-system namespace already contains a number
of infrastructural components that would benefit from having fluentd
configurations defined using the config-reloader and that would also
benefit from all the processing that the config-reloader can offer
on those configs, such as the "$labels" macro and so on.

This commit adds the possibility of configuring which is the namespace
that is used as an admin namespace through a flag of config-reloader
which defaults to "kube-system" for backwards compatibility. This way,
if the necessity of a different configuration arises, the config-
reloader can be installed to check for plugins and avoid validations of
configs from a different namespace.

Signed-off-by: Tommaso Pozzetti <tommypozzetti@hotmail.it>