Adds installCrd flag to helm chart values.
This solves the user experience problem of reading the documentation about
the CRD, running helm install, and not being able to see the CRD in the
api-resources.

Previously the CRD was only installed if the datasource was crd, or the
crdMigrationMode was turned on, which was a confusing experience for me
as a newcomer.

Authored-by: Adam Berlin <aberlin@vmware.com>
Signed-off-by: Adam Berlin <aberlin@vmware.com>